+++ ichnaea/async/schedule.py
@@ -129,13 +129,13 @@ CELERYBEAT_SCHEDULE = {
     },
     's3-delete-wifimeasures': {
         'task': 'ichnaea.backup.tasks.delete_wifimeasure_records',
-        'args': (100, 2, 300),
+        'args': (100, 2, 300, 10000),
         'schedule': crontab(hour=3, minute=17),
         'options': {'expires': 43200},
     },
     's3-delete-cellmeasures': {
         'task': 'ichnaea.backup.tasks.delete_cellmeasure_records',
-        'args': (100, 2, 300),
+        'args': (100, 2, 300, 10000),
         'schedule': crontab(hour=3, minute=27),
         'options': {'expires': 43200},
     },